* connection edges now contains read quality  * plot stage is an unsigned value  * other small improvements  * presentation from 14. 6.

+++ include/assembly.h
@@ -10,7 +10,7 @@
 typedef struct _PLOT_OPTIONS {
 	uint64_t PlotRefStart;
 	uint64_t PlotRefEnd;
-	int PlotStep;
+	uint16_t PlotStep;
 	uint16_t PlotFlags;
 } PLOT_OPTIONS, *PPLOT_OPTIONS;
